requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.3.0
blinker==1.7.0
click==8.1.7
duckdb==0.10.1
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
fsspec==2024.3.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
zipp==3.18.1
